Create an encryption key per secret
By using an encryption key per secret, there is no chance to unlock the secret even if you have the MASTER_KEY for the encryption/decryption. The encryption key is passed along with the secret, and is part of the url. Example: https://hemmelig.app/ENCRYPTION_KEY/SECRET_ID. TWhen encrypting, and decrypting, this happens: 32 length md5 hash md5(MASTER_KEY + ENCRYPTION_KEY), is used as the key.