security: prevent LFI if the disk adapter is being used
Hopefully this will prevent it. Have to dig deeper into this by going through som SecLists for LFI. Before this fix I was able to delete files on the filesystem such as /etc/passwd if the user was a root user. Was not able to view any other file than json files which had the same content structured as the secret files, though.